Work branch zhengtao 1

See merge request cese/student/software-systems/23-24/Group19!4

+++ 2-performance/src/config/mod.rs
@@ -40,7 +40,7 @@ pub enum GeneratorConfig {
 pub struct GeneralConfig {
     /// Very small float value.
     /// Two floats that are closer together than this value will be equal.
-    epsilon: f64,
+    epsilon: f32,
 
     /// Filename of the scene that will render
     scenename: String,
@@ -66,7 +66,7 @@ pub struct CameraConfig {
     height: usize,
 
     /// The field of view of the camera
-    fov: f64,
+    fov: f32,
 }
 
 impl Config {
